Login form with bcrypt
The username and password are stored in the component's state using the useState() hook, and the handleUsernameChange() and handlePasswordChange() methods are used to update the state when the user types in the input fields. The handleSubmit() method is called when the user submits the form, and it checks if the username and password are not empty and if they are correct. If the username and password are correct, it displays a success message. Otherwise, it displays an error message.